requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
Flask-Limiter==3.5.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-socketio==5.10.0
